Assert AutoChainRules{Int} for portable integer width.
On 32-bit Julia the Auto(:ChainRules, 1) config parameter is Int32,
so AutoChainRules{Int64} fails while AutoChainRules{Int} matches both.

+++ test/symbols.jl
@@ -1,7 +1,7 @@
 using ADTypes
 using Test
 
-@test ADTypes.Auto(:ChainRules, 1) isa AutoChainRules{Int64}
+@test ADTypes.Auto(:ChainRules, 1) isa AutoChainRules{Int}
 @test ADTypes.Auto(:Diffractor) isa AutoDiffractor
 @test ADTypes.Auto(:Enzyme) isa AutoEnzyme
 @test ADTypes.Auto(:FastDifferentiation) isa AutoFastDifferentiation
